requirement.txt
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
future==0.16.0
idna==2.6
networkx==2.1
nltk==3.2.5
oauthlib==2.0.7
pygraphviz==1.3.1
python-twitter==3.4.1
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
